requirements.txt
asgiref==3.7.2
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Django==4.2.11
duplicity==2.1.4
fasteners==0.18
Flask==3.0.2
gpg==1.18.0
gyp==0.1
httplib2==0.20.4
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
screen-resolution-extra==0.0.0
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
systemd-python==235
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
wadllib==1.3.6
Werkzeug==3.0.1
wheel==0.42.0
xdg==5
xkit==0.0.0
responses==0.23.1
pytest==7.4.0
pytest-django==4.5.2
coverage==7.2.7
django==4.2.11
requests>=2.31.0
python-nmap>=0.7.1
responses==0.23.1
pytest==7.4.0
pytest-django==4.5.2
coverage==7.2.7
dnspython>=2.4.2
# Add these lines to your requirements.txt
python-owasp-zap-v2.4>=0.0.20
docker>=6.1.2
requests>=2.31.0
urllib3>=2.0.7